[refac] #198 닉네임 중복검사 API의 책임을 user profile로 변경
