+++ hilingual-api/src/main/java/org/sopt/controller/userprofile/service/UserProfileService.java
@@ -1,7 +1,10 @@
 package org.sopt.controller.userprofile.service;
 
 import lombok.RequiredArgsConstructor;
+import org.sopt.controller.user.dto.NicknameAvailableRes;
+import org.sopt.controller.user.exception.UserSuccessCode;
 import org.sopt.controller.userprofile.dto.UserProfileReq;
+import org.sopt.dto.BaseResponseDto;
 import org.sopt.user.facade.UserFacade;
 import org.sopt.user.domain.User;
 import org.sopt.userprofile.domain.UserProfile;
@@ -15,6 +18,25 @@ import org.springframework.stereotype.Service;
 public class UserProfileService {
     private final UserFacade userFacade;
     private final UserProfileFacade userProfileFacade;
+
+    // TODO : 닉네임 중복 체크 아예 Custom Validator 로 빼자
+
+    private static final String NICKNAME_PATTERN = "^[가-힣ㄱ-ㅎㅏ-ㅣa-zA-Z0-9]+$";
+    private static final int MIN_NICKNAME_LENGTH = 2;
+    private static final int MAX_NICKNAME_LENGTH = 10;
+
+    public BaseResponseDto<NicknameAvailableRes> getNicknameAvailable(String nickname) {
+        if (!isValidFormat(nickname)) {
+            return unavailableNickname(UserSuccessCode.NICKNAME_SPECIAL_SYMBOLS);
+        }
+        if (!isValidLength(nickname)) {
+            return unavailableNickname(UserSuccessCode.NICKNAME_COUNT);
+        }
+        if (userFacade.isNicknameExists(nickname)) {
+            return unavailableNickname(UserSuccessCode.NICKNAME_DUPLICATED);
+        }
+        return availableNickname();
+    }
 
     public void save(Long userId, UserProfileReq userProfileReq) {
         // TODO : Custom error
@@ -31,4 +53,22 @@ public class UserProfileService {
         user.setIsCompleted(true);
         userFacade.save(user);
     }
+
+    private boolean isValidFormat(String nickname) {
+        return nickname.matches(NICKNAME_PATTERN);
+    }
+
+    private boolean isValidLength(String nickname) {
+        int length = nickname.length();
+        return length >= MIN_NICKNAME_LENGTH && length <= MAX_NICKNAME_LENGTH;
+    }
+
+    private BaseResponseDto<NicknameAvailableRes> availableNickname() {
+        return BaseResponseDto.success(UserSuccessCode.NICKNAME_AVAILABLE, new NicknameAvailableRes(true));
+    }
+
+    private BaseResponseDto<NicknameAvailableRes> unavailableNickname(UserSuccessCode code) {
+        return BaseResponseDto.success(code, new NicknameAvailableRes(false));
+    }
+
 }
